Moved Logger to Loggers
