+++ Blackboard/Core/Formula/Actions/TriggerGetter.cs
@@ -1,5 +1,6 @@
 ﻿using Blackboard.Core.Extensions;
 using Blackboard.Core.Inspect;
+using Blackboard.Core.Inspect.Loggers;
 using Blackboard.Core.Nodes.Collections;
 using Blackboard.Core.Nodes.Interfaces;
 using Blackboard.Core.Record;
