+++ Blackboard/Core/Formula/Formula.cs
@@ -1,6 +1,7 @@
 ﻿using Blackboard.Core.Extensions;
 using Blackboard.Core.Formula.Actions;
 using Blackboard.Core.Inspect;
+using Blackboard.Core.Inspect.Loggers;
 using System.Collections.Generic;
 using System.Linq;
