+++ Blackboard/Core/Inspect/Loggers/ConsoleLogger.cs
@@ -1,7 +1,7 @@
 ﻿using IO = System.IO;
 using S = System;
 
-namespace Blackboard.Core.Inspect;
+namespace Blackboard.Core.Inspect.Loggers;
 
 /// <summary>A logger that outputs to a console.</summary>
 sealed public class ConsoleLogger : Logger {
