+++ Blackboard/Core/Optimization/Rules/ConstantReduction.cs
@@ -1,5 +1,5 @@
 ﻿using Blackboard.Core.Extensions;
-using Blackboard.Core.Inspect;
+using Blackboard.Core.Inspect.Loggers;
 using Blackboard.Core.Nodes.Interfaces;
 
 namespace Blackboard.Core.Optimization.Rules;
