+++ BlackboardTests/Tools/FormulaExt.cs
@@ -2,6 +2,7 @@
 using Blackboard.Core.Formula;
 using Blackboard.Core.Formula.Actions;
 using Blackboard.Core.Inspect;
+using Blackboard.Core.Inspect.Loggers;
 using Blackboard.Core.Record;
 using System.Linq;
